% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htsk.R
\name{htsk}
\alias{htsk}
\title{Download a Hugging Face dataset and convert it into an mlr3 Task}
\usage{
htsk(repo_id, filename, target)
}
\arguments{
\item{repo_id}{Character string. Hugging Face dataset repository ID
(e.g. \code{"scikit-learn/iris"}).}

\item{filename}{Character string.
Name of the dataset file inside the repository.}

\item{target}{Character string. Name of the target column.}
}
\value{
An \code{mlr3} Task object (\code{TaskClassif} or \code{TaskRegr}).
}
\description{
This function downloads a dataset file from the Hugging Face Hub
and converts it into either a \code{TaskClassif} or \code{TaskRegr}
object from the \code{mlr3} package.
}
\details{
The task type is automatically determined based on the type of the
target column:
\itemize{
\item Factor/character → Classification
\item Numeric → Regression
}
}
\examples{
\dontrun{
task <- htsk(
  dataset = "scikit-learn/iris",
  config = "default",
  target = "Species"
)
task
}

}
